+++ models/src/anemoi/models/models/hierarchical.py
@@ -172,7 +172,7 @@ class AnemoiModelEncProcDecHierarchical(AnemoiModelEncProcDec):
         # Instantiation of model output bounding functions (e.g., to ensure outputs like TP are positive definite)
         self.boundings = nn.ModuleList(
             [
-                instantiate(cfg, name_to_index=self.data_indices.internal_model.output.name_to_index)
+                instantiate(cfg, name_to_index=self.data_indices.model.output.name_to_index)
                 for cfg in getattr(model_config.model, "bounding", [])
             ]
         )
