+++ training/src/anemoi/training/losses/scalers/variable.py
@@ -95,10 +95,10 @@ class GeneralVariableLossScaler(BaseVariableLossScaler):
         Retrieve the loss scaling for each variable from the config file.
         """
         variable_loss_scaling = (
-            np.ones((len(self.data_indices.internal_data.output.full),), dtype=np.float32) * self.weights.default
+            np.ones((len(self.data_indices.data.output.full),), dtype=np.float32) * self.weights.default
         )
 
-        for variable_name, idx in self.data_indices.internal_model.output.name_to_index.items():
+        for variable_name, idx in self.data_indices.model.output.name_to_index.items():
             _, variable_ref, _ = self.variable_metadata_extractor.get_group_and_level(variable_name)
             # Apply variable scaling by variable name
             # or base variable name (variable_ref: variable name without variable level)
