+++ training/tests/integration/aicon/test_cicd_aicon_04_icon-dream_medium.yaml
@@ -159,7 +159,6 @@ data:
   num_features: null
   format: zarr
   diagnostic: []
-  remapped:
   remapper:
     default: "none"
